add further 'evil' PDFs which crash SyncFunction library (used as one of the modes to calculate the pagecount of a PDF) --> exception the base constructor of the `AugmentedPdfLoadedDocument`
Incidentally, here you see a `documents/Z/` directory being added: that is my quick-hack to have the library facility detect new PDFs which I don't want to place in their proper slot manually. It's a little nasty, but it's a way to add PDFs in bulk to an existing library.