adjust the script generator for the move from `mujstest` to `bulktest` plus the introduction of metascripts (bulktest.bulktest is an example) where bulktest is allowed to invoke itself recursively.
The reason to do that would be the ability to easily run bulktests in the Visual Studio debugger or other environments with minimal parameterization (a setting that will be lost when you clean up the MSVC caches such as the /.vs/ directory when you *cleanup* before doing a major/full rebuild.

Regenerated the scripts.