added a ZIP collecting the curiously failing filenames' test set. As it says in the added README:
As an extra measure to ensure that `git` itself doesn't botch the job anywhere, we have included a ZIP file (`cannot-open-failure-collection.zip`) in this directory which contains all the files gathered here.
When in doubt of the veracity of git's reconstruction of the files on your system, delete the PDFs and extract them from the ZIP file.